+++ credentials-support/src/test/java/se/swedenconnect/security/credential/monitoring/DefaultCredentialTestFunctionTest.java
@@ -15,11 +15,15 @@
  */
 package se.swedenconnect.security.credential.monitoring;
 
+import jakarta.annotation.Nonnull;
+import lombok.NonNull;
 import org.cryptacular.io.ClassPathResource;
 import org.cryptacular.io.Resource;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import se.swedenconnect.security.credential.KeyStoreCredential;
+import se.swedenconnect.security.credential.PkiCredential;
 import se.swedenconnect.security.credential.ReloadablePkiCredential;
 import se.swedenconnect.security.credential.factory.KeyStoreFactory;
 import se.swedenconnect.security.credential.monitoring.DefaultCredentialMonitorBeanTest.TestCredential;
@@ -30,6 +34,8 @@ import java.security.KeyStore;
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
 import java.security.PrivateKey;
+import java.security.PublicKey;
+import java.util.Map;
 import java.util.function.Function;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -45,69 +51,82 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
  */
 public class DefaultCredentialTestFunctionTest {
 
-  /** Password for all keys. */
-  private static final char[] password = "secret".toCharArray();
-
-  /** RSA key */
-  private final ReloadablePkiCredential rsaCred;
-
-  /** DSA key */
-  private final ReloadablePkiCredential dsaCred;
-
-  /** EC key */
-  private final ReloadablePkiCredential ecCred;
-
-  public DefaultCredentialTestFunctionTest() throws Exception {
-    final Resource resource = new ClassPathResource("rsa-dsa-ec.jks");
-    try (final InputStream is = resource.getInputStream()) {
-      final KeyStore store = KeyStoreFactory.loadKeyStore(is, password, null, null);
-      this.rsaCred = new KeyStoreCredential(store, "rsa", password);
-      this.dsaCred = new KeyStoreCredential(store, "dsa", password);
-      this.ecCred = new KeyStoreCredential(store, "ec-nist", password);
+  private static ReloadablePkiCredential getCredential(final String alias) {
+    final char[] password = "secret".toCharArray();
+    try {
+      final Resource resource = new ClassPathResource("rsa-dsa-ec.jks");
+      try (final InputStream is = resource.getInputStream()) {
+        final KeyStore store = KeyStoreFactory.loadKeyStore(is, password, null, null);
+        return new KeyStoreCredential(store, alias, password);
+      }
+    }
+    catch (final Exception e) {
+      throw new RuntimeException(e);
     }
   }
 
   @Test
-  public void testCredentials() {
+  void testSigning() {
     final Function<ReloadablePkiCredential, Exception> func = new DefaultCredentialTestFunction();
 
-    Exception result = func.apply(this.rsaCred);
+    final ReloadablePkiCredential rsa = getCredential("rsa");
+    Exception result = func.apply(rsa);
     assertNull(result, "Test of RSA key was not successful");
 
-    result = func.apply(this.dsaCred);
+    rsa.getMetadata().setUsage(PkiCredential.Metadata.USAGE_SIGNING);
+    result = func.apply(rsa);
+    assertNull(result, "Test of RSA key was not successful");
+
+    final ReloadablePkiCredential dsa = getCredential("dsa");
+    result = func.apply(dsa);
     assertNull(result, "Test of DSA key was not successful");
 
-    result = func.apply(this.ecCred);
+    final ReloadablePkiCredential ec = getCredential("ec-nist");
+    ec.getMetadata().setUsage(PkiCredential.Metadata.USAGE_SIGNING);
+    result = func.apply(ec);
     assertNull(result, "Test of EC key was not successful");
   }
 
   @Test
-  public void testProviders() {
+  void testDecryption() {
+    final Function<ReloadablePkiCredential, Exception> func = new DefaultCredentialTestFunction();
+    final ReloadablePkiCredential rsa = getCredential("rsa");
+    rsa.getMetadata().setUsage(PkiCredential.Metadata.USAGE_ENCRYPTION);
+
+    final Exception result = func.apply(rsa);
+    assertNull(result, "Test of RSA key was not successful");
+  }
+
+  @Test
+  void testProviders() {
     final DefaultCredentialTestFunction func = new DefaultCredentialTestFunction();
 
+    final ReloadablePkiCredential rsa = getCredential("rsa");
+    final ReloadablePkiCredential dsa = getCredential("dsa");
+
     func.setProvider("SunRsaSign");
-    Exception result = func.apply(this.rsaCred);
+    Exception result = func.apply(rsa);
     assertNull(result, "Test of RSA key was not successful");
 
     // DSA should not work with SunRsaSign provider ...
-    result = func.apply(this.dsaCred);
+    result = func.apply(dsa);
     assertNotNull(result, "Expected NoSuchAlgorithmException result");
     assertTrue(result instanceof NoSuchAlgorithmException, "Expected NoSuchAlgorithmException exception");
 
     // This should work ...
     func.setProvider("SUN");
-    result = func.apply(this.dsaCred);
+    result = func.apply(dsa);
     assertNull(result, "Test of DSA key was not successful");
 
     // Non-existing provider ...
     func.setProvider("FooBar");
-    result = func.apply(this.rsaCred);
+    result = func.apply(rsa);
     assertNotNull(result, "Expected NoSuchProviderException result");
     assertTrue(result instanceof NoSuchProviderException, "Expected NoSuchProviderException exception");
   }
 
   @Test
-  public void testErrors() {
+  void testErrors() {
     final DefaultCredentialTestFunction func = new DefaultCredentialTestFunction();
 
     // NPE
@@ -117,23 +136,23 @@ public class DefaultCredentialTestFunctionTest {
 
     // Signing failed
     func.setRsaSignatureAlgorithm("SHA256withDSA");
-    result = func.apply(this.rsaCred);
+    result = func.apply(getCredential("rsa"));
     assertNotNull(result, "Expected KeyException result");
     assertTrue(result instanceof KeyException, "Expected KeyException exception");
 
     func.setDsaSignatureAlgorithm("SHA256withRSA");
-    result = func.apply(this.dsaCred);
+    result = func.apply(getCredential("dsa"));
     assertNotNull(result, "Expected KeyException result");
     assertTrue(result instanceof KeyException, "Expected KeyException exception");
 
     func.setEcSignatureAlgorithm("SHA256withRSA");
-    result = func.apply(this.ecCred);
+    result = func.apply(getCredential("ec-nist"));
     assertNotNull(result, "Expected KeyException result");
     assertTrue(result instanceof KeyException, "Expected KeyException exception");
   }
 
   @Test
-  public void testNoPrivateKey() {
+  void testNoPrivateKey() {
     final DefaultCredentialTestFunction func = new DefaultCredentialTestFunction();
     final TestCredential cred = new TestCredential("test");
 
@@ -142,13 +161,14 @@ public class DefaultCredentialTestFunctionTest {
   }
 
   @Test
-  public void testUnknownKeyAlgorithm() {
+  void testUnknownKeyAlgorithm() {
     final PrivateKey key = Mockito.mock(PrivateKey.class);
     Mockito.when(key.getAlgorithm()).thenReturn("UNKNOWN_ALGO");
 
     final ReloadablePkiCredential cred = Mockito.mock(ReloadablePkiCredential.class);
     Mockito.when(cred.getPrivateKey()).thenReturn(key);
     Mockito.when(cred.getName()).thenReturn("test");
+    Mockito.when(cred.getMetadata()).thenReturn(Map::of);
 
     final DefaultCredentialTestFunction func = new DefaultCredentialTestFunction();
     final Exception result = func.apply(cred);
@@ -156,19 +176,39 @@ public class DefaultCredentialTestFunctionTest {
   }
 
   @Test
-  public void testNullSettersAreIgnored() {
+  void testFallbackOnEncryption() {
+    final ReloadablePkiCredential rsa = getCredential("rsa");
+    final ReloadablePkiCredential cred = Mockito.mock(ReloadablePkiCredential.class);
+    Mockito.when(cred.getPrivateKey())
+        .thenReturn(rsa.getPrivateKey())
+        .thenThrow(new SecurityException("CKR_KEY_TYPE_INCONSISTENT"))
+        .thenReturn(rsa.getPrivateKey());
+    Mockito.when(cred.getName()).thenReturn("test");
+    Mockito.when(cred.getMetadata()).thenReturn(rsa.getMetadata());
+
+    // Assert that decrypt is called
+    final PublicKey pk = rsa.getPublicKey();
+    Mockito.when(cred.getPublicKey()).thenReturn(pk);
+
+    final DefaultCredentialTestFunction func = new DefaultCredentialTestFunction();
+    Assertions.assertNull(func.apply(cred));
+    Mockito.verify(cred, Mockito.times(1)).getPublicKey();
+  }
+
+  @Test
+  void testNullSettersAreIgnored() {
     final DefaultCredentialTestFunction func = new DefaultCredentialTestFunction();
     func.setDsaSignatureAlgorithm(null);
     func.setRsaSignatureAlgorithm(null);
     func.setEcSignatureAlgorithm(null);
 
-    Exception result = func.apply(this.rsaCred);
+    Exception result = func.apply(getCredential("rsa"));
     assertNull(result);
 
-    result = func.apply(this.dsaCred);
+    result = func.apply(getCredential("dsa"));
     assertNull(result);
 
-    result = func.apply(this.ecCred);
+    result = func.apply(getCredential("ec-nist"));
     assertNull(result);
   }
 }